Models/Assessment models/requirements.txt
# Core API
fastapi==0.116.1
starlette==0.47.2
uvicorn==0.30.6
python-multipart==0.0.12

# Config / Env
python-dotenv==1.0.1

# HTTP
requests==2.32.3

# LangChain (LLM integration)
langchain-core==0.3.21
langchain-openai==0.2.6

# Resume parsing
pdfplumber==0.11.4
python-docx==1.1.2

# Pydantic (ensure wheels; avoid building from source)
pydantic==2.11.7
pydantic-core==2.33.2
annotated-types==0.7.0
typing-extensions==4.12.2
